Shuffle most of the MemoryMap ugliness into a source file.
